% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_test.R
\name{mod_test}
\alias{mod_test}
\title{Modified paired samples t-test}
\usage{
mod_test(data = NULL, sig = 0.05)
}
\arguments{
\item{data}{list constructed by "simdat" function. Supplies data to apply
t-test on.}

\item{sig}{numeric. Supplies the significance level for the test.}
}
\value{
Returns the p-value of the performed test.
}
\description{
Performs a one-sided paired samples t-test modified according to Zimmerman
(2005; 2012).
}
\examples{
\donttest{
data <- simdat(npairs = 10, eff = 0, rho = -0.8)
test <- mod_test(data = data)
test
}

}
