% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_test.R
\name{paired_test}
\alias{paired_test}
\title{Paired t-test}
\usage{
paired_test(data = NULL, sig = 0.05)
}
\arguments{
\item{data}{list constructed by "simdat" function. Supplies data to apply
t-test on.}

\item{sig}{numeric. Supplies the significance level for the test.}
}
\value{
Returns the p-value of the performed test.
}
\description{
Performs a one-sided paired t-test.
}
\examples{
\donttest{
data <- simdat(npairs = 10, eff = 0, rho = -0.8)
test <- paired_test(data = data)
test
}

}
