% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powcorplot.R
\name{powcorplot}
\alias{powcorplot}
\title{Plotting of power curves I}
\usage{
powcorplot(
  gran = 100,
  numpairs = 10,
  effect = 0,
  diffdisplay = FALSE,
  mod = FALSE
)
}
\arguments{
\item{gran}{integer. Supplies the level of granularity, i.e. the number of
points for interpolation of the final curve.}

\item{numpairs}{integer. Supplies the number of pairs to be simulated.
Resulting in two samples of size "npairs".}

\item{effect}{numeric. Supplies the difference in means between both groups.
Can be interpreted as effect size based on the fact that standard normal
variates are the basis of simulation.}

\item{diffdisplay}{logical. If TRUE, plots will show difference between the
realised power or significance levels of paired and unpaired test. The
difference is calculated by subtracting the values of the unpaired test from
the paired tests values.}

\item{mod}{logical. If TRUE, plots will show the realised power or
significance levels of paired and  modified paired tests. The difference for
diffdisplay = TRUE is calculated by subtracting the values of the paired test
from the modified paired tests values.}
}
\value{
Returns plots of either the power curves(or significance levels)
of paired and unpaired t-test in relation to the underlying group correlations,
or alternatively the plotted difference between both curves.
}
\description{
Plotting of estimated power curves of unpaired and paired t-test in relation
to underlying group correlation.
In case of no actual mean difference, i.e. effect = 0, the curves represent
the observed rate of type I errors and should be interpreted as the
significance level of the tests.
}
\examples{
\donttest{
powcorplot(gran = 10, numpairs = 5, effect = 0.5)
}

}
