% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powr.R
\name{powr}
\alias{powr}
\title{Estimation of power for unpaired and paired t-test.}
\usage{
powr(npairs = NULL, eff = NULL, rho = NULL, sig = 0.05, rounds = 10000)
}
\arguments{
\item{npairs}{integer. Supplies the number of pairs to be simulated.
Resulting in two samples of size "npairs".}

\item{eff}{numeric. Supplies the difference in means between both groups.
Can be interpreted as effect size based on the fact that standard normal
variates are the basis of simulation.}

\item{rho}{numeric. Supplies the correlation coefficient between both groups.}

\item{sig}{numeric. Supplies the significance level for the test.
0.05 by default.}

\item{rounds}{integer. Supplies the number of simulation to base the power
estimation on. Set to 10000 by default.}
}
\value{
Returns the estimated power for both tests as a vector containing the
power of the unpaired t-test in the first place.
}
\description{
Estimates the power of paired and unpaired t-test in the case of normal data
for a given intra-pair correlation, significance level, sample size and
effect size.
}
\examples{
\donttest{
powertest <- powr(npairs = 10, eff = 0.5, rho = 0.1)
powertest
}

}
