% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samsizeplot.R
\name{samsizeplot}
\alias{samsizeplot}
\title{Plotting of power curves II}
\usage{
samsizeplot(upp = 100, corr = 0, effect = 0, mod = FALSE)
}
\arguments{
\item{upp}{integer. Supplies the upper limit of the sample sizes to be
simulated.}

\item{corr}{numeric. Supplies the underlying group correlation.}

\item{effect}{numeric. Supplies the difference in means between both groups.
Can be interpreted as effect size based on the fact that standard normal
variates are the basis of simulation.}

\item{mod}{logical. If TRUE, modified and paired t-test are compared.}
}
\value{
Returns plots of the power curves(or significance levels)
of paired and unpaired t-test in relation to the underlying sample sizes.
}
\description{
Plotting of estimated power curves of unpaired and paired t-test in relation
to underlying sample size.
In case of no actual mean difference, i.e. effect = 0, the curves represent
the observed rate of type I errors and should be interpreted as the
significance level of the tests.
}
\examples{
\donttest{
samsizeplot(upp = 100, corr = 0.3, effect = 0)
}

}
