% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulation of normal paired data}
\usage{
simdat(npairs = NULL, eff = NULL, rho = NULL)
}
\arguments{
\item{npairs}{integer. Supplies the number of pairs to be simulated.
Resulting in two samples of size "npairs".}

\item{eff}{numeric. Supplies the difference in means between both groups.
Can be interpreted as effect size based on the fact that standard normal
variates are the basis of simulation.}

\item{rho}{numeric. Supplies the correlation coefficient between both groups.}
}
\value{
A sample of paired data, i.e. two samples of size "npairs" with
theoretical correlation coefficient "rho". Realised empirical correlation
may vary and is not guaranteed to be an exact match of "rho".
}
\description{
Simulates samples of normally distributed paired data according to specified
sample size, intra-pair correlation and effect size of a given treatment.
}
\examples{
\donttest{
data <- simdat(npairs = 10, eff = 0, rho = -0.8)
cor(data$y_Tr, data$y_NTr)
}

}
